# Conflicts:
#	app/src/main/java/com/rozdoum/socialcomponents/adapters/PostsAdapter.java
#	app/src/main/java/com/rozdoum/socialcomponents/managers/DatabaseHelper.java
#	app/src/main/java/com/rozdoum/socialcomponents/model/Post.java
#	app/src/main/res/layout/post_item_list_view.xml

+++ app/src/main/java/com/rozdoum/socialcomponents/managers/DatabaseHelper.java
@@ -217,7 +217,7 @@ public class DatabaseHelper {
                 postRef.runTransaction(new Transaction.Handler() {
                     @Override
                     public Transaction.Result doTransaction(MutableData mutableData) {
-                        Integer currentValue = mutableData.getValue(Integer.class);
+                        Long currentValue = mutableData.getValue(Long.class);
                         if (currentValue == null) {
                             mutableData.setValue(0);
                         } else {
