+++ app/src/main/java/com/rozdoum/socialcomponents/model/Post.java
@@ -20,8 +20,8 @@ public class Post implements Serializable, LazyLoading {
     private long createdDate;
     private String imagePath;
     private String authorId;
-    private int commentsCount;
-    private int likesCount;
+    private long commentsCount;
+    private long likesCount;
     private ItemType itemType;
 
     public Post() {
@@ -81,19 +81,19 @@ public class Post implements Serializable, LazyLoading {
         this.createdDate = createdDate;
     }
 
-    public int getCommentsCount() {
+    public long getCommentsCount() {
         return commentsCount;
     }
 
-    public void setCommentsCount(int commentsCount) {
+    public void setCommentsCount(long commentsCount) {
         this.commentsCount = commentsCount;
     }
 
-    public int getLikesCount() {
+    public long getLikesCount() {
         return likesCount;
     }
 
-    public void setLikesCount(int likesCount) {
+    public void setLikesCount(long likesCount) {
         this.likesCount = likesCount;
     }
